requirements.txt
cycler==0.11.0
kafka-python==2.0.2
kiwisolver==1.3.2
matplotlib==3.4.3
numpy==1.21.3
Pillow==8.4.0
pyparsing==3.0.4
python-dateutil==2.8.2
six==1.16.0
